+++ index.html
@@ -21,15 +21,16 @@
     content="Shopping Time is an e-commerce website that has at least 150 women's clothing products with different sizes and color choices">
   <meta name="twitter:image"
     content="https://github.com/catherineisonline/shopping-time/raw/main/src/assets/images/project-preview.webp?raw=true">
-  <link rel="icon" href="%PUBLIC_URL%/favicon.ico">
   <meta name="viewport" content="width=device-width, initial-scale=1">
   <meta name="theme-color" content="#000000">
-  <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png">
+  <link rel="icon" href="/favicon.ico">
+  <link rel="apple-touch-icon" href="/logo192.png">
   <title>Shopping Time</title>
 </head>
 
 <body>
   <div id="root"></div>
+  <script type="module" src="/src/index.jsx"></script>
 </body>
 
 </html>
