+++ src/components/attributes/SelectedAttributes.jsx
