+++ src/components/cart-overlay/CartIcon.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useRef, useState } from "react";
 import EmptyCart from "../../assets/images/cart-icon.png";
-import CartOverlay from "./CartOverlay.js";
-import { ResetLocation } from "../../helpers/ResetLocation";
+import CartOverlay from "./CartOverlay.jsx";
+import { ResetLocation } from "../../helpers/ResetLocation.jsx";
 
 const CartIcon = ({ totalPayment,
   cartItems,
