+++ src/index.jsx
@@ -1,15 +1,14 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import App from "./App.js";
+import App from "./App.jsx";
 import "./core-ui/styles.css";
 import "./core-ui/hovers.css";
 import "./core-ui/responsive.css";
-import "./routes/single-product/single-product.css";
-
+import "./routes/single-product/single-product.css"; 
 import reportWebVitals from "./reportWebVitals";
+
 const root = ReactDOM.createRoot(document.getElementById("root"));
-root.render(
-  <App />
-);
+root.render(<App />);
+
 
 reportWebVitals();
