+++ src/routes/all-products/Product.jsx
